iprt/RTPathGlob.cpp: VC++ 14.1 build fix. bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@83758 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/path/RTPathGlob.cpp
@@ -1367,7 +1367,7 @@ static int rtPathGlobParse(PRTPATHGLOB pGlob, const char *pszPattern, PRTPATHPAR
             else
                 AssertMsgFailedReturn(("'%.*s' is not supported yet\n", pszComp, pParsed->aComps[0].cch),
                                       VERR_PATH_MATCH_FEATURE_NOT_IMPLEMENTED);
-            pGlob->offFirstPath = (uint32_t)RTPathEnsureTrailingSeparator(pGlob->szPath, sizeof(pGlob->szPath));
+            pGlob->offFirstPath = (uint16_t)RTPathEnsureTrailingSeparator(pGlob->szPath, sizeof(pGlob->szPath));
             pGlob->iFirstComp   = iComp = 1;
         }
     }
